reorganize code for clarity
Reduce duplicated code in nullModelTestPrep by combining the
mixed model and unrelated hetResid cases. Code path now depends
on whether cholSigmaInv is a scalar or a matrix. The prior case
of treating a diagonal matrix differently by extracting the
diagonal elements is no longer necessary, since operations with
a Diagonal Matrix class in the Matrix package are just as fast.