allow weights in GRanges object used to create iterator
Instead of looking for weights in variantData, use currentVariants
which will get them from either variantData or the GRanges/
GRangesList object used to create the iterator. This allows
provision of weights without using the variantData slot, and
allows different weights for the same variant depending on what
unit it's in.