if null model doesn't contain sample.id, use appropriate samples
Although not recommended, a user might create a null model with
a data.frame object that does not contain the sample.id. If the
gdsobj is from SeqArray, assume that the sample filter is set as
desired to match the null model and do not change it. If the
gdsobj is from GWASTools, use the row names of the model matrix
to index the samples.