reorder phenotypes to match cov.mat in null model
Since the covariance matrix may be block diagonal, we should not
reorder it. Instead, reorder the data frame to match cov.mat.
fitNullModel for AnnotatedDataFrame now calls the data.frame
method, and createDesignMatrix is a function, not a method, that
operates on a data.frame only. This makes it easier to deal with
all sample.id operations in one place. The sample.id argument no
longer specifies the order of samples, only which ones to include.