Write a single function instead of an internal function with wrapper
At this point, only one type of input is planned: a null model and
a matrix of genotypes. Instead of having an internal function and
a series of wrapper functions, just handle everything in this one
function. It can be split in the future if we need to fit a joint
model using different classes of inputs.