Remove elements in fit element from null model object
This includes resid.marginal, resid.conditional, outcome, etc.
This breaks most of the assoc testing code, so those tests will
fail after this commit. They'll be fixed in future commits. In
addition, some tests related to the null model are failing because
the outcome element is expected to be a 1-column matrix with a
column name. Now it's a vector in the fit data frame, so we need
to decide what to do about those tests. Should we add a new element
to the null model indicating the name of the outcome (eg as a
string)?