Remove sample.id element from null model object
I had forgotten to comment out a line in the fitNullModel code that
added a "sample.id" element to the null model object in addition to
adding it to the "fit" data frame element. Remove this line and
update code and tests such that they pass using the sample.id column
in the fit data frame instead of the now-removed "sample.id" element.