Fix tests for glmm null model
The test data for the null model didn't work properly for glmm
models; all variance components went to zero, so the function fit
a regular logistic model instead of one with variance components.
Fix the test data so it works properly for glmm models. Add a second
test that explicitly tests running a regular glm model when the
variance components go to zero.