Add comments about moving model string updates to a wrapper
There are the two functions that update a null model to use a
different outcome or to add a conditional variant. These function
are awaiting wrappers. It would make more sense in the long run to
update the model string in the wrapper instead of in the main
functions, but those wrappers don't exist yet. Add a comment
indicating that those lines should be moved once the wrapper exists.