Do not calculate resid.conditional for WLS models
Remove the lines calculating and returning resid.conditional for
WLS models. When updating older null models to the new format, only
return resid.conditional if the model is a mixed model (by removing
it when it's not). Finally, update the old WLS null model in the
testdata; it didn't appear to be correct for some reason.