use BiocParallel for assocTest functions
Use the bpiterate function in BiocParallel to control iteration in
association tests (single and aggregate). The iterator function
relies on iterateFilter and returns all genotype/variant data
necessary for each block. The function to process genotype data,
including computing allele frequency and doing the association test,
is shared between the two methods for each function, so there is
less duplicated code, and class-specific code is concentrated in
the iterator functions.