add error handling to assoc tests using bpiterate
bpiterator returns errors as elements of a list with valid
results, leading to uninformative error messages from rbindlist
when it attempts to bind an error message to a data frame. Add a
utility function to check the results of bpiterate for errors,
then stop the function with the original error message, as well as
on which iteration it occurred.