bug fix in SKAT-O if a test returns pval=NA
Allow selecting an alternate minimum p-value even if some SKAT-O
tests return NA. Also, previously if the pval was NA, opt.rho
would attempt to index a missing value and return a zero-length
vector, so the resulting output could not be converted to a data
frame and thew an error. Now we have an explicit check for
missing values.