add support for SeqVarListIterator in metaPrepScores
Names for group IDs in SeqVarListIterator objects may be taken
from the names of the GRangesList object defining iterations.
If that is NULL, it is not straightforward to construct names
since variants in each group are not sequential nor even required
to be on the same chromosome. Instead, number groups with
sequential integers.