Update fitNullModel code to require all samples in cov.mat
Previously, sample selection was confusing:
- if sample.id was passed, cov.mat also had to be subset to that
  specific set of samples.
- if sample.id was not passed, the samples in cov.mat were used to
  determine the sample set. For example, if cov.mat only had 5
  samples, the null model would only include those 5 samples.
This meant that cov.mat often defined the sample set (without any
messages to the user) instead of the primary required input (the
AnnotatedDataFrame). Change the sample selection such that the
sample set is determined by 1) sample.id, if passed and 2) the
set of samples in the AnnotatedDataFrame, if sample.id is not
passed. This is clearer and easier to document for users. To help
with those, split the .checkSampleId function into two functions
that check the sample set in the AnnotatedDataFrame and in the
cov.mat argument separately.